+++ tests/testthat/test-tbl_svysummary.R
@@ -22,8 +22,8 @@ test_that("tbl_svysummary creates output without error/warning (no by var)", {
 
 test_that("tbl_svysummary creates output without error/warning (with by var)", {
   statistics <- list(
-    all_continuous() ~ "{median} {mean} {sd} {var} {min} {max} {sum} {p25} {p42} {p75} {p89} {mean.std.error}",
-    all_categorical() ~ "{n} {N} {p} | {n_unweighted} {N_unweighted} {p_unweighted}"
+    all_continuous() ~ "{median} {mean} {sd} {var} {min} {max} {sum} {p25} {p42} {p75} {p89} {mean.std.error} {deff}",
+    all_categorical() ~ "{n} {N} {p} | {n_unweighted} {N_unweighted} {p_unweighted} {p.std.error} {deff}"
   )
   expect_snapshot(
     tbl_svysummary(dc_light, by = both, statistic = statistics) %>% render_as_html()